change elastic index
new index is { photo_id: Int, taxon_id: [Int], embedding: [Float] }
where taxon_id is the list of all taxon ids for each observation
that a photo belongs to, as well as the ancestry for each of
those observations. so we're doing some stuff with pandas now.
also, honor the batch size in config
also, allow user to set a cap from an export. helpful since
with human exclusion we don't know how many photos will actually
be ingested from a dataset